[Refactor]: move cartesian2polar to data.py
- finalize fourier.py by removing old fuunctions
